% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sar.R
\name{calculate_sar_indices}
\alias{calculate_sar_indices}
\title{Calculer les indices radar SAR}
\usage{
calculate_sar_indices(sar_raster)
}
\arguments{
\item{sar_raster}{SpatRaster avec bandes VV et VH (en dB)}
}
\value{
SpatRaster avec ratio VV/VH et RVI
}
\description{
Calculer les indices radar SAR
}
