% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_models.R
\name{classify_crop_stage}
\alias{classify_crop_stage}
\title{Classifier les stades de croissance des cultures}
\usage{
classify_crop_stage(sar_features, labels, method = c("rf", "svm"))
}
\arguments{
\item{sar_features}{Data frame avec colonnes VV, VH, RVI, ratio_VV_VH}

\item{labels}{Vecteur de stades observés ("début", "développement", "maturité")}

\item{method}{"rf" ou "svm"}
}
\value{
Modèle de classification entraîné
}
\description{
Classifier les stades de croissance des cultures
}
