% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sar.R
\name{clean_sar_data}
\alias{clean_sar_data}
\title{Nettoyer les données SAR (filtre anti-speckle + valeurs aberrantes)}
\usage{
clean_sar_data(sar_raster, filter_size = 3)
}
\arguments{
\item{sar_raster}{SpatRaster}

\item{filter_size}{Taille du filtre médian (défaut : 3)}
}
\value{
SpatRaster nettoyé
}
\description{
Nettoyer les données SAR (filtre anti-speckle + valeurs aberrantes)
}
