% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sar.R
\name{convert_to_db}
\alias{convert_to_db}
\title{Convertir les valeurs SAR en décibels}
\usage{
convert_to_db(sar_raster)
}
\arguments{
\item{sar_raster}{SpatRaster avec bandes VV et/ou VH (valeurs linéaires)}
}
\value{
SpatRaster en dB
}
\description{
Convertir les valeurs SAR en décibels
}
