% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sar_data.R
\name{download_sentinel1}
\alias{download_sentinel1}
\title{Télécharger des images Sentinel-1 via l'API STAC Copernicus}
\usage{
download_sentinel1(bbox, date_start, date_end, output_dir = "data-raw/")
}
\arguments{
\item{bbox}{Vecteur numérique c(xmin, ymin, xmax, ymax) en WGS84}

\item{date_start}{Date de début "YYYY-MM-DD"}

\item{date_end}{Date de fin "YYYY-MM-DD"}

\item{output_dir}{Répertoire de sortie}
}
\value{
Chemins vers les fichiers téléchargés
}
\description{
Télécharger des images Sentinel-1 via l'API STAC Copernicus
}
