% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_models.R
\name{estimate_soil_moisture}
\alias{estimate_soil_moisture}
\title{Estimer l'humidité du sol par régression ou Random Forest}
\usage{
estimate_soil_moisture(sar_features, field_data, method = c("lm", "rf"))
}
\arguments{
\item{sar_features}{Data frame avec colonnes VV, VH, ratio_VV_VH}

\item{field_data}{Data frame terrain avec colonne soil_moisture}

\item{method}{"lm" ou "rf"}
}
\value{
Modèle entraîné (lm ou randomForest)
}
\description{
Estimer l'humidité du sol par régression ou Random Forest
}
