% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilgrids.R
\name{get_soilgrids_point}
\alias{get_soilgrids_point}
\title{Récupérer l'humidité du sol via l'API SoilGrids (ISRIC)}
\usage{
get_soilgrids_point(lon, lat, depth = "0-5cm", property = "wv0033")
}
\arguments{
\item{lon}{Longitude du point}

\item{lat}{Latitude du point}

\item{depth}{Profondeur : "0-5cm", "5-15cm", "15-30cm"}

\item{property}{Propriété : "wv0033" (humidité vol.), "clay", "sand", "silt"}
}
\value{
Data frame avec les valeurs de la propriété
}
\description{
Récupérer l'humidité du sol via l'API SoilGrids (ISRIC)
}
