% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_model_validation}
\alias{plot_model_validation}
\title{Graphique de validation : observé vs prédit}
\usage{
plot_model_validation(eval_result)
}
\arguments{
\item{eval_result}{Résultat de evaluate_model()}
}
\value{
Graphique ggplot2
}
\description{
Graphique de validation : observé vs prédit
}
