% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_sar_map}
\alias{plot_sar_map}
\title{Cartographier les bandes SAR ou l'humidité du sol}
\usage{
plot_sar_map(raster, band = "VV_dB", title = "Carte SAR", output_path = NULL)
}
\arguments{
\item{raster}{SpatRaster à visualiser}

\item{band}{Nom de la bande à afficher ("VV_dB", "VH_dB", "RVI", etc.)}

\item{title}{Titre de la carte}

\item{output_path}{Chemin d'export PNG/PDF (optionnel)}
}
\value{
Objet tmap
}
\description{
Cartographier les bandes SAR ou l'humidité du sol
}
