% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_sar_timeseries}
\alias{plot_sar_timeseries}
\title{Tracer l'évolution temporelle des signaux SAR}
\usage{
plot_sar_timeseries(ts_data, interactive = TRUE)
}
\arguments{
\item{ts_data}{Data frame avec colonnes : date, VV, VH, culture}

\item{interactive}{Si TRUE, retourne un graphique plotly interactif}
}
\value{
Graphique ggplot2 ou plotly
}
\description{
Tracer l'évolution temporelle des signaux SAR
}
